add getAllApplicants query and test, and clean up seeder helper functions
