+++ src/server/auth.ts
@@ -19,6 +19,7 @@ import { env } from "~/env";
 import { type Database, db } from "~/server/db";
 import { createTable, users } from "~/server/db/schema";
 import { UserSeeder } from "./db/seed/userSeeder";
+import { PgInsertValue } from "drizzle-orm/pg-core";
 
 /**
  * Module augmentation for `next-auth` types. Allows us to add custom properties to the `session`
@@ -123,6 +124,19 @@ export const getServerAuthSession = (ctx: {
 
 export async function mockSession(db: Database): Promise<Session> {
   const user = new UserSeeder().createRandom();
+  await db.insert(users).values(user).returning();
+
+  return {
+    user,
+    expires: new Date(Date.now() + 10).toISOString(),
+  };
+}
+
+export async function mockOrganizerSession(db: Database): Promise<Session> {
+  const user = {
+    ...new UserSeeder().createRandom(),
+    type: "organizer",
+  } as const satisfies PgInsertValue<typeof users>;
   await db.insert(users).values(user).returning();
 
   return {
